Refactor Room component and update BottomBar integration
- Replaced ButtonsPaneModal with BottomBarBtns and integrated AudioSelectModal into the Room component.
- Updated BottomBar to accept bottomBarBtns as a prop for improved modularity.
- Adjusted LoginScreen to include paddingBottom for better layout.
- Enhanced TermsOfUseModal styling and structure for improved visibility and usability.
- Removed unused ButtonsPaneModal component to streamline the codebase.
- Added isAudioSelectOpen state management in zustand for audio selection modal visibility control.
