+++ src/bottomBar/BottomBarBtns.js
@@ -1,0 +1,40 @@
+import * as React from 'react';
+import { StyleSheet, View } from 'react-native';
+import { baseStyles } from '../constants';
+import { AudioModeBtn } from './AudioModeBtn';
+import { CammuteBtn } from './CammuteBtn';
+import { MoreBtn } from './MoreBtn';
+import { MuteBtn } from './MuteBtn';
+import { QuestionBtn } from './QuestionBtn';
+
+export const BottomBarBtns = () => {
+  return (
+    <View style={[styles.container, baseStyles.panelBackground]}>
+      <MuteBtn />
+      <CammuteBtn />
+      <QuestionBtn />
+      <AudioModeBtn />
+      <View style={styles.devider}></View>
+      <MoreBtn />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    minWidth: 340,
+    maxWidth: '100%',
+    flexDirection: 'row',
+    paddingVertical: 16,
+    paddingHorizontal: 12,
+    borderRadius: 32,
+  },
+  devider: {
+    width: 1,
+    backgroundColor: '#333',
+    marginLeft: 4,
+    marginRight: 4,
+    marginTop: 16,
+    marginBottom: 16,
+  },
+});
