+++ src/libs/connection-monitor.js
@@ -2,7 +2,7 @@ import NetInfo from '@react-native-community/netinfo';
 import BackgroundTimer from 'react-native-background-timer';
 import kc from '../auth/keycloak';
 import logger from '../services/logger';
-import { sleep } from '../tools';
+import { rejectTimeoutPromise, sleep } from '../tools';
 import { useInRoomStore } from '../zustand/inRoom';
 import { useInitsStore } from '../zustand/inits';
 import { useSettingsStore } from '../zustand/settings';
