+++ src/zustand/shidur.js
@@ -131,6 +131,11 @@ export const useShidurStore = create((set, get) => ({
   isMuted: false,
   shidurWIP: false,
   cleanWIP: false,
+  isAudioSelectOpen: false,
+
+  setIsAudioSelectOpen: (isAudioSelectOpen = !get().isAudioSelectOpen) => {
+    set({ isAudioSelectOpen });
+  },
 
   setIsMuted: (isMuted = !get().isMuted) => {
     if (!get().isPlay) {
